Big refactor that stores operator LLVM modules at the RuntimePlan level.
Prior to this change, each snippet of JIT compiled code in a Trecul dataflow was stored in its own LLVM module; thus a single plan could easily carry 10's to 100's of modules.   This has bothered me for quite some time and I finally got around to fixing it.    Firstly, some refactoring of the RecordType* classes needed to happen so that the serializable reference to an LLVM function could be separated from the serializable module and the executable function pointer.   Previously references were also executable (the IQL*Module classes).   The references are now Trecul*Reference classes, the executable objects are Trecul*Runtime classes and there is an TreculModule to represent the module.  Trecul*References are not executable objects.  A TreculModule instance now exists at the RuntimePlan level and is de/serialized with the plan.   Next, each operator had to be modified to be given access to a per-plan CodeGenerationContext and store and serialize Trecul*References for each JIT compiled function used.   Moreover, we now need a post-deserialization step to lookup Trecul*References contained in each RuntimeOperatorType in the plan's TreculModule and get corresponding Trecul*Runtime instances.  Public interfaces to the RecordType* and IQL*Module classes are unchanged but the implementations of these classes are now using the TreculModule based classes.