Dusted off the original MPI parallel execution code and got it working with current Trecul.
This is should be treated as experimental at this point and it has only been tested with mpich-4.2.3 (in fact it is known not to work yet with the Open MPI distribution on my Ubuntu workstation).  In the process of doing this work infrastructure for intra-process parallelism (multiple threads, partitioners and collectors) has been added as well.   Infrastructure for parallel/multi-partition unit tests has been added and those will always run single process mode and will run multi-process if MPI is configured.